requirements.txt
fastapi
uvicorn
pandas
chromadb
langchain
langchain-chroma
langchain-huggingface
sentence-transformers
streamlit
pytest
httpx
pydantic-settings
